pyproject.toml
[tool.poetry]
name = "excursionist"
version = "0.1.0"
description = ""
authors = ["daniarlert <arlertdaniel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.9.0"
scrapy-user-agents = "^0.1.1"
scrapy-playwright = "^0.0.26"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.15"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
